agregando detalles de funciones

+++ src/pages/Home.jsx
@@ -10,6 +10,8 @@ import { ToastContainer } from "react-toastify";
 import { getImagesFromLocalStorage } from "../util/imageReload";
 import { handleShowImageDetails } from "../util/showImage";
 import { showImageDetailsForm } from "../util/showImageDetailsForm";
+import Swal from "sweetalert2";
+
 
 export function Home() {
   const [images, setImages] = useState([]);
@@ -18,10 +20,30 @@ export function Home() {
   getImagesFromLocalStorage(setImages);
 
   const handleDeleteImage = (public_id) => {
-    deleteImage(public_id);
-    setImages((prevImages) =>
-      prevImages.filter((image) => image.public_id !== public_id)
-    );
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        
+        deleteImage(public_id);
+        setImages((prevImages) =>
+        prevImages.filter((image) => image.public_id !== public_id)
+      );
+        Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
+      }
+    })
+    
+   
   };
 
   return (
